refactor: add sanity checker to common as the function
