+++ docs/apis/common.md
@@ -7,6 +7,8 @@
 
 ::: t4_devkit.common.io
 
+::: t4_devkit.common.sanity
+
 ::: t4_devkit.common.serialize
 
 ::: t4_devkit.common.timestamp
