+++ t4_devkit/cli/sanity.py
@@ -1,25 +1,14 @@
 from __future__ import annotations
 
-import re
-import warnings
 from pathlib import Path
 
 import typer
-from attrs import define
 from tabulate import tabulate
 from tqdm import tqdm
 
-from t4_devkit import Tier4
+from t4_devkit.common.sanity import DBException, sanity_check
 
 from .version import version_callback
-
-
-@define
-class DBException:
-    dataset_id: str
-    version: str | None
-    message: str
-
 
 cli = typer.Typer(
     name="t4sanity",
@@ -29,31 +18,15 @@ cli = typer.Typer(
 )
 
 
-def _check_sanity(db_parent: str) -> list[DBException]:
+def _run_sanity_check(db_parent: str, *, include_warning: bool = False) -> list[DBException]:
     exceptions: list[DBException] = []
 
     db_dirs: list[Path] = Path(db_parent).glob("*")
-    version_pattern = re.compile(r".*/\d+$")
 
-    for db_root in tqdm(db_dirs, desc=">>> Sanity checking..."):
-        versions = [d.name for d in db_root.iterdir() if version_pattern.match(str(d))]
-        if versions:
-            version = sorted(versions)[-1]
-            data_root = db_root.joinpath(version).as_posix()
-        else:
-            version = None
-            data_root = db_root.as_posix()
-
-        try:
-            _ = Tier4("annotation", data_root=data_root, verbose=False)
-        except Exception as e:
-            exceptions.append(
-                DBException(
-                    dataset_id=db_root.name,
-                    version=version,
-                    message=str(e),
-                )
-            )
+    for db_root in tqdm(db_dirs, desc=">>>Sanity checking..."):
+        result = sanity_check(db_root, include_warning=include_warning)
+        if result:
+            exceptions.append(result)
     return exceptions
 
 
@@ -72,12 +45,7 @@ def main(
         False, "-iw", "--include-warning", help="Indicates whether to report any warnings."
     ),
 ) -> None:
-    with warnings.catch_warnings():
-        if include_warning:
-            warnings.filterwarnings("error")
-        else:
-            warnings.filterwarnings("ignore")
-        exceptions = _check_sanity(db_parent)
+    exceptions = _run_sanity_check(db_parent, include_warning=include_warning)
 
     headers = ["DatasetID", "Version", "Message"]
     table = [[e.dataset_id, e.version, e.message] for e in exceptions]
